Replace `mainReady`
