+++ apps/meteor/app/ui-utils/client/index.ts
@@ -1,5 +1,4 @@
 export { messageBox } from './lib/messageBox';
 export { LegacyRoomManager } from './lib/LegacyRoomManager';
 export { upsertMessage, RoomHistoryManager } from './lib/RoomHistoryManager';
-export { mainReady } from './lib/mainReady';
 export { MessageTypes, MessageType } from '../lib/MessageTypes';
