--- apps/meteor/app/ui-utils/client/lib/mainReady.ts
@@ -1,3 +1,0 @@
-import { ReactiveVar } from 'meteor/reactive-var';
-
-export const mainReady = new ReactiveVar(false);
