+++ apps/meteor/client/views/root/MainLayout/EmbeddedPreload.tsx
@@ -1,23 +1,12 @@
-import { useUserId } from '@rocket.chat/ui-contexts';
 import type { ReactElement, ReactNode } from 'react';
 import { useEffect } from 'react';
 
-import { mainReady } from '../../../../app/ui-utils/client';
-import { RoomsCachedStore, SubscriptionsCachedStore, PublicSettingsCachedStore } from '../../../cachedStores';
-import { useUserDataSyncReady } from '../../../lib/userData';
+import { RoomsCachedStore, SubscriptionsCachedStore } from '../../../cachedStores';
 import PageLoading from '../PageLoading';
+import { useMainReady } from '../hooks/useMainReady';
 
 const EmbeddedPreload = ({ children }: { children: ReactNode }): ReactElement => {
-	const uid = useUserId();
-	const subscriptionsReady = SubscriptionsCachedStore.useReady();
-	const settingsReady = PublicSettingsCachedStore.useReady();
-	const userDataReady = useUserDataSyncReady();
-
-	const ready = !uid || (userDataReady && subscriptionsReady && settingsReady);
-
-	useEffect(() => {
-		mainReady.set(ready);
-	}, [ready]);
+	const ready = useMainReady();
 
 	useEffect(() => {
 		SubscriptionsCachedStore.setReady(true);
